Initialize authorization mechanism
The authorization is based on the idea that the remote proxy will send
us a serie of headers and we trust those headers.

There are two types of roles in the system: super admin (defined by SUPERADMIN_USERNAME
env variable) and regular users.
The superadmin can do (a priori) everything and has a nice flash message
when they exercise their power.
Regular users can only act on themselves (and for now create new users).

For now we only implement authorization on user management. Future
patches will extend this to all objects.

Change-Id: Id3f5b29dfdbe05740ac0b0245449527b77dfe42b